+++ Controller/Webhook/Webhook.php
@@ -4,6 +4,7 @@ namespace Dibs\EasyCheckout\Controller\Webhook;
 
 use Dibs\EasyCheckout\Model\Checkout as DibsCheckout;
 use Dibs\EasyCheckout\Model\CheckoutContext as DibsCheckoutContext;
+use Dibs\EasyCheckout\Model\Client\DTO\UpdatePaymentReference;
 use Magento\Framework\App\Action\HttpPostActionInterface;
 use Magento\Framework\App\CsrfAwareActionInterface;
 use Magento\Framework\App\Request\InvalidRequestException;
@@ -86,6 +87,8 @@ abstract class Webhook implements HttpPostActionInterface, CsrfAwareActionInterf
     //
 
     public function __construct(
+        \Dibs\EasyCheckout\Helper\Data $helper,
+        \Dibs\EasyCheckout\Model\Client\Api\Payment $paymentApi,
         RequestInterface $request,
         DibsCheckout $dibsCheckout,
         DibsCheckoutContext $dibsCheckoutContext,
@@ -95,6 +98,8 @@ abstract class Webhook implements HttpPostActionInterface, CsrfAwareActionInterf
         string $successComment,
         bool $sendEmail
     ) {
+        $this->helper = $helper;
+        $this->paymentApi = $paymentApi;
         $this->request = $request;
         $this->dibsCheckout = $dibsCheckout;
         $this->dibsCheckoutContext = $dibsCheckoutContext;
@@ -154,6 +159,22 @@ abstract class Webhook implements HttpPostActionInterface, CsrfAwareActionInterf
             $this->dibsCheckoutContext->getOrderRepository()->save($this->order);
             $this->paymentRepo->save($this->order->getPayment());
             $this->logInfo("Order is updated successfully");
+            $orderId = $this->order->getIncrementId();
+            $payment = $this->order->getPayment();
+            if($payment->getMethod() == "dibseasycheckout"){
+                $paymentId = $this->order->getDibsPaymentId();
+                $reference = new UpdatePaymentReference();
+                $reference->setReference($this->order->getIncrementId());
+                $reference->setCheckoutUrl($this->helper->getCheckoutUrl());
+                if ($this->helper->getCheckoutFlow() === "HostedPaymentPage") {
+                    $payment = $this->paymentApi->getPayment($paymentId);
+                    $checkoutUrl = $payment->getCheckoutUrl();
+                    $reference->setCheckoutUrl($checkoutUrl);
+                }
+                
+                $this->logInfo('from the webhook ' . $this->order->getIncrementId() );
+                $this->paymentApi->UpdatePaymentReference($reference, $paymentId);
+            }
         } catch (\Exception $e) {
             $this->logError("Could not update order");
             $this->logError("Error message: {$e->getMessage()}");